.github/scripts/requirements.txt
requests>=2.31.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pyyaml>=6.0

